pyproject.toml
[project]
name = "fuel-supply-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
